common: move string functions to separate file
